requirements.txt
asgiref==3.2.10
astroid==2.6.5
attrs==20.2.0
certifi==2020.6.20
chardet==4.0.0
coverage==5.2.1
dj-rest-auth==2.1.4
Django==2.2.24
django-cors-headers==3.5.0
django-crontab==0.7.1
django-elasticsearch-dsl==6.4.2
django-elasticsearch-dsl-drf==0.20.8
django-mysql==3.8.1
django-nine==0.2.3
djangorestframework==3.11.2
ElasticMock==1.5.1
elasticsearch==6.8.1
elasticsearch-dsl==6.4.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.0.1
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mock==4.0.2
more-itertools==8.5.0
mysqlclient==2.0.1
packaging==20.4
pluggy==0.13.1
py==1.10.0
pylint==2.9.5
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
responses==0.12.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
uWSGI==2.0.19.1
wrapt==1.12.1
